+++ experiments/kitchen/plan2explore/plan2explore_intrinsic_raw_actions.py
@@ -7,7 +7,7 @@ from rlkit.launchers.launcher_util import run_experiment
 from rlkit.torch.model_based.dreamer.experiments.experiment_utils import (
     preprocess_variant,
 )
-from rlkit.torch.model_based.plan2explore.experiments.kitchen_plan2explore import (
+from rlkit.torch.model_based.plan2explore.experiments.raps_plan2explore_experiment import (
     experiment,
 )
