+++ experiments/metaworld/dreamer/dreamer_v2_single_task_raw_actions.py
@@ -4,7 +4,7 @@ import subprocess
 
 import rlkit.util.hyperparameter as hyp
 from rlkit.launchers.launcher_util import run_experiment
-from rlkit.torch.model_based.dreamer.experiments.kitchen_dreamer import experiment
+from rlkit.torch.model_based.dreamer.experiments.raps_experiment import experiment
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
