+++ experiments/skill_learn/dreamer_v2_single_task_low_level_primitives_kitchen_raps_params.py
@@ -7,9 +7,7 @@ from rlkit.launchers.launcher_util import run_experiment
 from rlkit.torch.model_based.dreamer.experiments.experiment_utils import (
     preprocess_variant,
 )
-from rlkit.torch.model_based.dreamer.experiments.low_level_primitives_exp import (
-    experiment,
-)
+from rlkit.torch.model_based.dreamer.experiments.ll_raps_experiment import experiment
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
