+++ experiments/skill_learn/train_wm.py
@@ -4,7 +4,9 @@ import subprocess
 
 import rlkit.util.hyperparameter as hyp
 from rlkit.launchers.launcher_util import run_experiment
-from rlkit.torch.model_based.dreamer.experiments.wm_prim_exp import experiment
+from rlkit.torch.model_based.dreamer.experiments.world_model_training_experiment import (
+    experiment,
+)
 
 if __name__ == "__main__":
     parser = argparse.ArgumentParser()
